uplink(deps): Bump uplink-sys version
Bump the uplink-sys version to the latest published one.

It has required to adapt some values to uszie because before they were
u64, hence, we don't need any casting.

Because the new version brings an error code that was not mapped because
of a bug in the past, we had to adjust the test to check for this new
error code rather than being an internal error, which is used when it's
an internal one, but also if an error code doesn't match any of the
registered ones.

+++ uplink/src/object.rs
@@ -239,9 +239,8 @@ impl std::io::Read for Download {
             let bp = buf.as_mut_ptr();
             // SAFETY: we trust the FFI of dealing with a correct `UplinkDownload` instance and an
             // allocated buffer.
-            let read_res = unsafe {
-                ulksys::uplink_download_read(self.inner.download, bp.cast(), buf.len() as u64)
-            };
+            let read_res =
+                unsafe { ulksys::uplink_download_read(self.inner.download, bp.cast(), buf.len()) };
 
             if let Some(err) = Error::new_uplink(read_res.error) {
                 // According to the Uplink C bindings version that we are targeting v1.7.0 all the
